fixed interrupted uploads

+++ src/it/ksuploader/main/RequestHandler.java
@@ -8,6 +8,8 @@ import java.io.IOException;
 import java.io.RandomAccessFile;
 import java.nio.channels.FileChannel;
 import java.nio.channels.SocketChannel;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 import javax.imageio.ImageIO;
 
@@ -141,7 +143,6 @@ public class RequestHandler implements Runnable {
 			FileChannel fileChannel;
 			aFile = new RandomAccessFile(fileName, "rw");
 			fileChannel = aFile.getChannel();
-			try {
 
 				long fileLength = dis.readLong();
 				System.out.println("File length: " + fileLength);
@@ -153,25 +154,14 @@ public class RequestHandler implements Runnable {
 				Thread.sleep(1000);
 				fileChannel.close();
 				System.out.println("End of file reached, closing channel");
+                if(fileLength !=  new File(fileName).length()){
+                    System.out.println("File invalid, deleting...");
+                    new File(fileName).delete();
+                }
 
-			} catch (IOException | InterruptedException e) {
-				try {
-					e.printStackTrace();
-					aFile.close();
-					fileChannel.close();
-					new File(fileName).delete();
-					System.out.println("Transfer cancelled.");
-					if (System.getProperty("os.name").toLowerCase().indexOf("win") >= 0)
-						Runtime.getRuntime().exec("del /f /q " + fileName);
-					System.out.println("Transfer cancelled.");
-				} catch (IOException ex) {
-					ex.printStackTrace();
-				}
-			}
-
-		} catch (FileNotFoundException ex) {
-			ex.printStackTrace();
-		}
+		} catch (IOException | InterruptedException ex) {
+            ex.printStackTrace();
+        }
 
 	}
